+++ app/(city)/[city]/(category)/[category]/page.tsx
@@ -1,7 +1,7 @@
 import { COMMON_DOMAIN, COMMON_TITLE } from '@/shared';
 import { Metadata, ResolvingMetadata } from 'next/types';
 import { FirmsPage } from './FirmsPage';
-import { getCategory, getFirms } from './api';
+import { getCategories, getCategory, getCities, getFirms } from './api';
 
 export interface CategoryPageProps {
   params: { city: string; category: string };
@@ -45,6 +45,17 @@ export default async function Page({ params, searchParams }: CategoryPageProps) 
 
   const category = await getCategory(cityAbbr);
   const firms = await getFirms(cityAbbr, categoryAbbr, firmsPage, 10);
+  const cities = await getCities();
+  const categories = await getCategories(1, 10);
 
-  return <FirmsPage cityAbbr={cityAbbr} categoryAbbr={categoryAbbr} category={category} firms={firms} />;
+  return (
+    <FirmsPage
+      cityAbbr={cityAbbr}
+      categoryAbbr={categoryAbbr}
+      category={category}
+      firms={firms}
+      cities={cities}
+      categories={categories}
+    />
+  );
 }
