+++ app/(city)/[city]/page.tsx
@@ -1,5 +1,5 @@
 import { CategoriesPage } from './CategoriesPage';
-import { getCategories, getCities, getCity } from './api';
+import { getCategories, getCities } from './api';
 
 export interface CityPageProps {
   params: {
@@ -11,8 +11,8 @@ export interface CityPageProps {
 export default async function Page({ params }: CityPageProps) {
   const cityId = params.city ?? '';
 
-  const city = await getCity(cityId);
+  const cities = await getCities();
   const categories = await getCategories(1, 10);
 
-  return <CategoriesPage cityId={cityId} categories={categories} />;
+  return <CategoriesPage cityId={cityId} cities={cities} categories={categories} />;
 }
