+++ app/(footer)/about/page.tsx
@@ -1,38 +1,28 @@
-'use client';
-import { Curve } from '@/features';
-import { COMMON_DOMAIN, COMMON_TITLE } from '@/shared';
-import { CommonHeader, Footer, Nav, Section } from '@/widgets';
+import { COMMON_DOMAIN, COMMON_HOST, COMMON_TITLE } from '@/shared';
+import { Metadata } from 'next';
+import { AboutPage } from './AboutPage';
+import { getCategories, getCities } from './api';
 
-export default function Page() {
-  return (
-    <div>
-      <Curve>
-        <Nav />
-        <Section>
-          <CommonHeader
-            title={`${COMMON_TITLE} - Сервис, который помогает развиваться локальному бизнесу в России и СНГ`}
-          />
-          <div className="w-full flex flex-col items-center gap-4 min-h-[500px]">
-            <div className="container min-h-[500px] w-full flex flex-col gap-4 px-8 py-10 overflow-hidden bg-white shadow-2xl rounded-xl dark:bg-gray-800">
-              <h2 className="text-lg dark:text-blue-400 text-blue-400">Что мы делаем?</h2>
-              <p>
-                Мы помогаем бизнесу из разных сфер — от медицинских центров и банкетных залов до ремонта телефонов и
-                шиномонтажа
-              </p>
-              <h2 className="text-lg dark:text-blue-400 text-blue-400">Миссия</h2>
-              <p>
-                Помочь бизнесу заниматься тем, для чего он создан: кормить, лечить, делать людей красивыми — а мы
-                возьмем на себя остальные задачи
-              </p>
-              <h2 className="text-lg dark:text-blue-400 text-blue-400">Цифры и факты</h2>
-              <p>{`Каждый месяц ${COMMON_DOMAIN} посещают 1 млн человек, мы активно расширяем партнерскую сеть`}</p>
-            </div>
-          </div>
-          <div className="flex w-full">
-            <Footer />
-          </div>
-        </Section>
-      </Curve>
-    </div>
-  );
+export async function generateMetadata(): Promise<Metadata> {
+  return {
+    title: `О нас ${COMMON_TITLE} - рейтинг кафе, баров, фастфудов, цены, фото, телефоны, адреса, отзывы`,
+    description: `Выбор лучших услуг: рестораны, салоны красоты, медицина и многое другое на ${COMMON_DOMAIN}. Фотографии, отзывы, акции, скидки, фильтры для поиска.`,
+    alternates: { canonical: `${COMMON_HOST}/about` },
+    keywords: ['отзывы', ' рейтинг', ' рестораны', ' салоны красоты', ' автосервисы', ' медицина'],
+    openGraph: {
+      title: `О нас ${COMMON_TITLE} - рейтинг кафе, баров, фастфудов, цены, фото, телефоны, адреса, отзывы`,
+      description: `Выбор лучших услуг: рестораны, салоны красоты, медицина и многое другое на ${COMMON_DOMAIN}. Фотографии, отзывы, акции, скидки, фильтры для поиска.`,
+      url: `${COMMON_HOST}/about`,
+      siteName: `${COMMON_DOMAIN}`,
+      locale: 'ru_RU',
+      type: 'website',
+    },
+  };
+}
+
+export default async function Page() {
+  const cities = await getCities();
+  const categories = await getCategories(1, 10);
+
+  return <AboutPage cities={cities} categories={categories} />;
 }
