+++ app/(footer)/contacts/page.tsx
@@ -1,33 +1,27 @@
-'use client';
-import { Curve } from '@/features';
-import { COMMON_TITLE } from '@/shared';
-import { CommonHeader, Footer, Nav, Section } from '@/widgets';
+import { COMMON_DOMAIN, COMMON_HOST, COMMON_TITLE } from '@/shared';
+import { Metadata } from 'next';
+import { ContactsPage } from './ContactsPage';
+import { getCategories, getCities } from './api';
 
-export default function Page() {
-  return (
-    <div>
-      <Curve>
-        <Nav />
-        <Section>
-          <CommonHeader title={`Контакты сайта каталога организаций ${COMMON_TITLE}`} />
-          <div className="w-full flex flex-col items-center gap-4 min-h-[500px]">
-            <div className="container min-h-[500px] w-full flex flex-col gap-4 px-8 py-10 overflow-hidden bg-white shadow-2xl rounded-xl dark:bg-gray-800">
-              <p>Набережные Челны, пр. Московский, 140А (52/03А) деловой центр Форт Диалог, этаж 2, офис 200</p>
-              <p>
-                {`Телефон для связи: `}
-                <a href="tel:+79112395458">+7 (911) 239-54-58</a>
-              </p>
-              <p>
-                {`Почта для запросов на сотрудничество: `}
-                <a href="mailto:kir.dontsov@gmail.com">kir.dontsov@gmail.com</a>
-              </p>
-            </div>
-          </div>
-          <div className="flex w-full">
-            <Footer />
-          </div>
-        </Section>
-      </Curve>
-    </div>
-  );
+export async function generateMetadata(): Promise<Metadata> {
+  return {
+    title: `Контакты ${COMMON_TITLE} - рейтинг кафе, баров, фастфудов, цены, фото, телефоны, адреса, отзывы`,
+    description: `Выбор лучших услуг: рестораны, салоны красоты, медицина и многое другое на ${COMMON_DOMAIN}. Фотографии, отзывы, акции, скидки, фильтры для поиска.`,
+    alternates: { canonical: `${COMMON_HOST}/caontacts` },
+    keywords: ['отзывы', ' рейтинг', ' рестораны', ' салоны красоты', ' автосервисы', ' медицина'],
+    openGraph: {
+      title: `Контакты ${COMMON_TITLE} - рейтинг кафе, баров, фастфудов, цены, фото, телефоны, адреса, отзывы`,
+      description: `Выбор лучших услуг: рестораны, салоны красоты, медицина и многое другое на ${COMMON_DOMAIN}. Фотографии, отзывы, акции, скидки, фильтры для поиска.`,
+      url: `${COMMON_HOST}/caontacts`,
+      siteName: `${COMMON_DOMAIN}`,
+      locale: 'ru_RU',
+      type: 'website',
+    },
+  };
+}
+
+export default async function Page() {
+  const cities = await getCities();
+  const categories = await getCategories(1, 10);
+  return <ContactsPage cities={cities} categories={categories} />;
 }
