+++ app/(root)/categories/CategoriesPage.tsx
@@ -1,6 +1,7 @@
 'use client';
 import { $cityError, Category } from '@/api';
 import { CategoriesGateProvider, CategoriesList, CitiesGateProvider, CityIdGateProvider, Curve } from '@/features';
+import { CommonNavProps } from '@/shared';
 import { CommonHeader, Footer, Nav, Section } from '@/widgets';
 import { useUnit } from 'effector-react';
 import { notFound } from 'next/navigation';
@@ -8,11 +9,10 @@ import { FC } from 'react';
 
 export interface CategoriesPageProps {
   cityId: string;
-  categories: Category[] | null;
 }
 
 /** Список категорий внутри города */
-export const CategoriesPage: FC<CategoriesPageProps> = ({ cityId, categories }) => {
+export const CategoriesPage: FC<CategoriesPageProps & CommonNavProps> = ({ cityId, cities, categories }) => {
   const cityError = useUnit($cityError);
   if (cityError) {
     notFound();
@@ -22,7 +22,7 @@ export const CategoriesPage: FC<CategoriesPageProps> = ({ cityId, categories }) 
       <CityIdGateProvider cityId={cityId}>
         <CategoriesGateProvider>
           <Curve>
-            <Nav />
+            <Nav cities={cities} categories={categories} />
             <Section>
               <CommonHeader title="Категории" subTitle="раздел" />
               <div className="container flex flex-col gap-4 items-center mb-auto">
